refactor: Split ChatPrompt into compact and full components, keeping logic on ChatPrompt
